fix: rearrange crates
