+++ crates/future/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "future"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+default = ["full"]
+full = ["metrics"]
+metrics = ["dep:metrics"]
+
+[dependencies]
+metrics = { path = "../metrics", optional = true }
+tokio = { version = "1", default-features = false, features = ["rt-multi-thread", "time", "macros"] }
+tokio-util = { version = "0.7", default-features = false }
+pin-project = "1"
+thiserror = "1.0"
