+++ crates/utils/Cargo.toml
@@ -2,13 +2,3 @@
 name = "utils"
 version = "0.1.0"
 edition = "2021"
-
-[dependencies]
-tokio = { version = "1", default-features = false, features = ["rt-multi-thread", "time", "macros"] }
-tokio-util = { version = "0.7", default-features = false }
-pin-project = "1"
-thiserror = "1.0"
-prometheus = "0.13"
-opentelemetry = { version = "0.18", features = ["metrics", "rt-tokio"] }
-opentelemetry-prometheus = "0.11"
-once_cell = "1.17"
