+++ crates/utils/src/lib.rs
@@ -1,3 +1,3 @@
+pub use collections::*;
+
 pub mod collections;
-pub mod futures;
-pub mod metrics;
