+++ src/lib.rs
@@ -1,2 +1,6 @@
+#[cfg(feature = "future")]
+pub use future;
+#[cfg(feature = "metrics")]
+pub use metrics;
 #[cfg(feature = "utils")]
 pub use utils;
